Add some more determinism to rhyme choices
previously the rhyme options were picked at random and then if after
several attempts nothing worked it would give up. now instead of picking
at random between all rhyme options, it filters it out to the groups
that are relevant. for example, the [🔬🔭] group will not be used
for a rhyme group with /xx, because none of the options match